Move edit series logic to hook
