+++ client/web/src/insights/pages/creation/search-insight/components/form-series/FormSeries.module.scss
@@ -11,19 +11,3 @@
         }
     }
 }
-
-.form-series-card {
-    &:hover {
-        box-shadow: 0 0 0 0.1875rem var(--primary);
-        border-color: transparent;
-    }
-
-    &__color {
-        width: 1rem;
-        height: 1rem;
-        align-self: center;
-        flex-grow: 0;
-        border-radius: 50%;
-        background-color: currentColor;
-    }
-}
