+++ client/web/src/insights/pages/creation/search-insight/components/form-series/FormSeries.tsx
@@ -1,29 +1,54 @@
 import classnames from 'classnames'
-import React, { ReactElement } from 'react'
+import React from 'react'
 
 import { DataSeries } from '../../../../../core/backend/types'
 import { FormSeriesInput } from '../form-series-input/FormSeriesInput'
 
+import { SeriesCard } from './components/series-card/SeriesCard';
 import styles from './FormSeries.module.scss'
 
 export interface FormSeriesProps {
-    /** Controlled value (series - chart lines) for series input component. */
+    /**
+     * Controlled value (series - chart lines) for series input component.
+     * */
     series?: DataSeries[]
 
+    /**
+     * Edit series used below for rendering edit series form. Element of
+     * this array has undefined value when there are no series to edit
+     * and has DataSeries value when user activated edit for some series.
+     * */
     editSeries: (DataSeries | undefined)[]
 
-    /** Change handler. */
-    // onChange: (series: DataSeries[]) => void
-
     /**
-     * Live change series while user typing in active series form.
-     * Used by consumers for getting latest values for live preview chart.
+     * Live change series handler while user typing in active series form.
+     * Used by consumers to get latest values from series inputs and pass
+     * them tp live preview chart.
      * */
     onLiveChange: (liveSeries: DataSeries, isValid: boolean, index: number) => void
 
-    onEditSeriesRequest: (openedCardIndex: number) => void
+    /**
+     * Handler that runs every time user clicked edit on particular
+     * series card.
+     * */
+    onEditSeriesRequest: (editSeriesIndex: number) => void
+
+    /**
+     * Handler that runs every time use clicked commit (done) in
+     * series edit form.
+     * */
     onEditSeriesCommit: (seriesIndex: number, editedSeries: DataSeries) => void
+
+    /**
+     * Handler that runs every time use canceled (click cancel) in
+     * series edit form.
+     * */
     onEditSeriesCancel: (closedCardIndex: number) => void
+
+    /**
+     * Handler that runs every time use removed (click remove) in
+     * series card.
+     * */
     onSeriesRemove: (removedSeriesIndex: number) => void
 }
 
@@ -85,58 +110,5 @@ export const FormSeries: React.FunctionComponent<FormSeriesProps> = props => {
                 + Add another data series
             </button>
         </ul>
-    )
-}
-
-interface SeriesCardProps {
-    /** Name of series. */
-    name: string
-    /** Query value of series. */
-    query: string
-    /** Color value of series. */
-    stroke: string
-    /** Custom class name for root button element. */
-    className?: string
-    /** Edit handler. */
-    onEdit?: () => void
-    /** Remove handler. */
-    onRemove?: () => void
-}
-
-/**
- * Renders series card component, visual list item of series (name, color, query)
- * */
-function SeriesCard(props: SeriesCardProps): ReactElement {
-    const { name, query, stroke: color, className, onEdit, onRemove } = props
-
-    return (
-        <li
-            aria-label={`Edit button for ${name} data series`}
-            className={classnames(styles.formSeriesCard, className, 'card d-flex flex-row p-3')}
-        >
-            <div className="flex-grow-1 d-flex flex-column align-items-start">
-
-                <div className='d-flex align-items-center mb-1 '>
-                    {/* eslint-disable-next-line react/forbid-dom-props */}
-                    <div style={{ color }} className={styles.formSeriesCardColor} />
-                    <span className="ml-1 font-weight-bold">{name}</span>
-                </div>
-
-                <span className="mb-0 text-muted">{query}</span>
-            </div>
-
-            <div className='d-flex align-items-center'>
-
-                <button
-                    type="button"
-                    onClick={onEdit}
-                    className='border-0 btn btn-outline-primary'>Edit</button>
-
-                <button
-                    type="button"
-                    onClick={onRemove}
-                    className='border-0 btn btn-outline-danger ml-1'>Remove</button>
-            </div>
-        </li>
     )
 }
