+++ client/web/src/insights/pages/creation/search-insight/components/search-insight-creation-content/hooks/use-editable-series.ts
@@ -1,0 +1,155 @@
+import { useMemo, useState } from 'react';
+
+import { useFieldAPI } from '../../../../../../components/form/hooks/useField';
+import { DataSeries } from '../../../../../../core/backend/types';
+import { useDistinctValue } from '../../../../../../hooks/use-distinct-value';
+import { CreateInsightFormFields } from '../../../types';
+import { DEFAULT_ACTIVE_COLOR } from '../../form-color-input/FormColorInput';
+
+const createDefaultEditSeries = (series = defaultEditSeries, valid = false): EditDataSeries => ({
+    ...series,
+    valid,
+})
+
+const defaultEditSeries = {
+    name: '',
+    query: '',
+    stroke: DEFAULT_ACTIVE_COLOR,
+}
+
+interface EditDataSeries extends DataSeries {
+    valid: boolean;
+}
+
+export interface UseEditableSeriesProps {
+    series: useFieldAPI<CreateInsightFormFields['series']>
+}
+
+export interface UseEditableSeriesAPI {
+    /**
+     * Edit series array used below for rendering series edit form.
+     * In case of some element has undefined value we're showing
+     * series card with data instead of form.
+     * */
+    editSeries: (CreateInsightFormFields['series'][number] | undefined)[],
+
+    /**
+     * Latest valid values of series.
+     * */
+    liveSeries: DataSeries[],
+
+    /**
+     * Handler to listen latest values of particular sereis form.
+     * */
+    listen: (liveSeries: DataSeries, valid: boolean, index: number) => void
+
+    /**
+     * Handlers for CRUD operations over sereis.
+     * */
+    editRequest: (index: number) => void
+    editCommit: (index: number, editedSeries: DataSeries) => void
+    cancelEdit: (index: number) => void
+    deleteSeries: (index: number) => void
+}
+
+/**
+ * Implementation of CRUD operation over insight series. Used in form to manage
+ * edit, delete, add, and cancel series forms.
+ * */
+export function useEditableSeries(props: UseEditableSeriesProps): UseEditableSeriesAPI {
+    const { series } = props;
+
+    const [editSeries, setEditSeries] = useState<(EditDataSeries | undefined)[]>(() => {
+        const hasSeries = series.input.value.length;
+
+        if (hasSeries) {
+            return series.input.value.map(() => undefined);
+        }
+
+        // If we in creation mode we should show first series editor in a first
+        // render.
+        return [createDefaultEditSeries()]
+    })
+
+    const liveSeries = useDistinctValue(
+        useMemo<DataSeries[]>(
+            () => editSeries
+                .map((editSeries, index) => {
+                    if (editSeries) {
+                        const { valid, ...series } = editSeries
+                        return valid ? series : undefined
+                    }
+
+                    return series.meta.value[index];
+                })
+                // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+                // @ts-ignore
+                .filter<DataSeries>(series => !!series),
+            [series, editSeries]
+        )
+    );
+
+    const handleSeriesLiveChange = (liveSeries: DataSeries, valid: boolean, index: number): void => {
+        const newEditSeries = [...editSeries];
+
+        newEditSeries[index] = { ...liveSeries, valid }
+
+        setEditSeries(newEditSeries)
+    }
+
+    const handleEditSeriesRequest = (index: number): void => {
+        const newEditSeries = [...editSeries];
+
+        newEditSeries[index] = series.meta.value[index]
+            ? createDefaultEditSeries(series.meta.value[index], true)
+            : createDefaultEditSeries()
+
+        setEditSeries(newEditSeries)
+    }
+
+    const handleEditSeriesCancel = (index: number): void => {
+        const newEditSeries = [...editSeries];
+
+        newEditSeries[index] = undefined
+        setEditSeries(newEditSeries)
+    }
+
+    const handleEditSeriesCommit = (index: number, editedSeries: DataSeries): void => {
+        const newEditedSeries = [...editSeries];
+        const newSeries = [
+            ...series.input.value.slice(0, index),
+            editedSeries,
+            ...series.input.value.slice(index + 1),
+        ]
+
+        // Remove series from edited cards
+        newEditedSeries[index] = undefined
+
+        setEditSeries(newEditedSeries)
+        series.input.onChange(newSeries)
+    }
+
+    const handleRemoveSeries = (index: number): void => {
+        const newSeries = [
+            ...series.input.value.slice(0, index),
+            ...series.input.value.slice(index + 1),
+        ]
+        const newEditedSereis = [
+            ...editSeries.slice(0, index),
+            ...editSeries.slice(index + 1),
+        ]
+
+        setEditSeries(newEditedSereis)
+        series.input.onChange(newSeries)
+    }
+
+    return {
+        liveSeries,
+        editSeries,
+        listen: handleSeriesLiveChange,
+        editRequest: handleEditSeriesRequest,
+        editCommit: handleEditSeriesCommit,
+        cancelEdit: handleEditSeriesCancel,
+        deleteSeries: handleRemoveSeries,
+    }
+}
